requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
django-allauth==0.44.0
django-filter==2.4.0
django-timezone-field==4.1.1
django-widget-tweaks==1.4.8
filelock==3.0.12
future==0.18.2
html5lib==1.1
idna==2.10
MouseInfo==0.1.3
oauthlib==3.1.0
Pillow==8.0.1
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
PyJWT==2.0.0
PyMsgBox==1.0.9
PyPDF2==1.26.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-bidi==0.4.2
python3-openid==3.2.0
PyTweening==1.0.3
pytz==2020.4
reportlab==3.5.68
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.3.1
webencodings==0.5.1
xhtml2pdf==0.2.5
